pyproject.toml
[tool.poetry]
authors = ["Rishabh Mittal <mittalrishabh181@gmail.com>"]
description = ""
name = "binance-trader"
version = "0.1.0"

[tool.poetry.dependencies]
numpy = "^1.23.2"
pandas = "^1.4.4"
python = "^3.10"
python-binance = "^1.0.16"
matplotlib = "^3.5.3"

[tool.poetry.dev-dependencies]
black = "^22.8.0"
ipykernel = "^6.15.2"
jupyterlab-code-formatter = "^1.5.3"
pytest = "^7.1.2"
jupyterlab = "^3.4.5"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]
